week6-ML/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
imbalanced-learn
missingno
pyyaml
scipy